release-2.5.6: Fix some compiler warnings

+++ include/read_alignment.h
@@ -25,8 +25,8 @@
 #include <boost/lexical_cast.hpp>
 using boost::lexical_cast;
 
-#include <string>
 #include <stdexcept>
+#include <string>
 #include <vector>
 
 #include "htslib/hts.h"
@@ -84,7 +84,7 @@ struct Align {
       return "chr-1";
     }
 
-    if (chromNum >= refVec.size()) {
+    if (chromNum >= (int)refVec.size()) {
       throw std::out_of_range(
           "[DecodeChrom ERROR] Input chromosme index: " +
           lexical_cast<std::string>(chromNum) + " but there are only " +
