+++ src/json_output.cc
@@ -28,8 +28,8 @@
 #include <string>
 #include <vector>
 
-#include "third_party/json/json.hpp"
 #include <boost/algorithm/string/join.hpp>
+#include "third_party/json/json.hpp"
 
 #include "common/parameters.h"
 #include "common/repeat_spec.h"
@@ -90,7 +90,7 @@ void WriteJson(const Parameters &parameters,
              region_findings.offtarget_irr_counts.size());
       auto &offtarget_section =
           results_json[region_findings.region_id]["OffTargetRegionIrrCounts"];
-      for (int i = 0; i != repeat_spec.offtarget_regions.size(); ++i) {
+      for (int i = 0; i != (int)repeat_spec.offtarget_regions.size(); ++i) {
         offtarget_section[repeat_spec.offtarget_regions[i].ToString()] =
             region_findings.offtarget_irr_counts[i];
       }
